Moved performance manager
